Chore: refactor sections
Move code out of provider.py, which was getting too big.
