+++ tests/unit/curated_recommendations/fixtures.py
@@ -11,21 +11,43 @@ from merino.curated_recommendations.corpus_backends.protocol import Topic
 from merino.curated_recommendations.layouts import layout_4_medium
 
 
-def generate_recommendations(item_ids: list[str]) -> list[CuratedRecommendation]:
-    """Create dummy recommendations."""
+def generate_recommendations(
+    length: int | None = None, item_ids: list[str] | None = None, time_sensitive_count: int | None = None
+) -> list[CuratedRecommendation]:
+    """Create dummy recommendations for the tests below.
+
+    @param length: Optionally, the number of items to generate. `length` or `item_ids` must be set.
+    @param item_ids: Optionally, a list of item ids.
+    @param time_sensitive_count: the number of items to make time-sensitive.
+        If None (the default), then half of the recommendations will be time-sensitive.
+    @return: A list of curated recommendations
+    """
     recs = []
-    for item_id in item_ids:
+
+    if item_ids is not None:
+        length = len(item_ids)
+    elif length is not None:
+        item_ids = [f"id:{i}" for i in range(length)]
+
+    # If time_sensitive_count is not provided, default to half the length
+    if time_sensitive_count is None:
+        time_sensitive_count = length // 2
+
+    # Randomly choose indices that will be time-sensitive
+    time_sensitive_indices = random.sample(range(length), time_sensitive_count)
+
+    for i, item_id in enumerate(item_ids):
         rec = CuratedRecommendation(
-            corpusItemId=str(uuid.uuid4()),
+            corpusItemId=item_id,
             tileId=MIN_TILE_ID + random.randint(0, 101),
-            receivedRank=random.randint(0, 101),
-            scheduledCorpusItemId=item_id,
+            receivedRank=i,
+            scheduledCorpusItemId=str(uuid.uuid4()),
             url=HttpUrl("https://littlelarry.com/"),
             title="little larry",
             excerpt="is failing english",
             topic=random.choice(list(Topic)),
             publisher="cohens",
-            isTimeSensitive=False,
+            isTimeSensitive=i in time_sensitive_indices,
             imageUrl=HttpUrl("https://placehold.co/600x400/"),
             iconUrl=None,
         )
